requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.172.0
requests==2.32.4
pydantic==2.11.6
